requirements.txt
numpy>=1.20.1
pandas>=1.1.3
tqdm>=4.46.0
matplotlib>=3.3.2
seaborn>=0.11.0
notebook>=6.1.4
ipywidgets>=7.5.1
scikit-learn>=0.23.2
scikit-image>=0.17.2
albumentations>=0.5.2
black>=20.8b1
torchviz>=0.0.1
pypatconsole>=0.0.20
